Refactor: extract transpilation utility
